requirements.txt
playsound==1.2.2
textblob==0.15.3
instaloader==4.5.5
SpeechRecognition==3.8.1
demoji==0.4.0
Pillow==8.0.1
